Minor code improvements
  -- related to "external disk for memory data support"

+++ wffweb/src/main/java/com/webfirmframework/wffweb/server/page/ExternalDriveByteArrayQueue.java
@@ -220,6 +220,10 @@ class ExternalDriveByteArrayQueue implements Queue<byte[]> {
 		long rId;
 		while ((rId = readId.get()) < writeId.get()) {
 			final long newReadId = rId + 1L;
+			if (writeIdInProgressStates.get(newReadId) != null) {
+				// writing is inprogress so break
+				break;
+			}
 			if (readId.compareAndSet(rId, newReadId)) {
 				deleteByReadId(newReadId);
 			}
